# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusOura API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserStudyMode = (typeof UserStudyMode)[keyof typeof UserStudyMode];

export const UserStudyMode = {
  light: "light",
  night: "night",
} as const;

export interface User {
  id: string;
  displayName: string;
  email: string;
  studyMode: UserStudyMode;
  notificationsEnabled: boolean;
  createdAt: string;
}

export type UpdateUserBodyStudyMode =
  (typeof UpdateUserBodyStudyMode)[keyof typeof UpdateUserBodyStudyMode];

export const UpdateUserBodyStudyMode = {
  light: "light",
  night: "night",
} as const;

export interface UpdateUserBody {
  displayName?: string;
  studyMode?: UpdateUserBodyStudyMode;
  notificationsEnabled?: boolean;
}

export interface Subject {
  id: string;
  name: string;
  accentColor: string;
  plantId?: string;
  totalFocusMinutes: number;
  sessionCount: number;
  createdAt: string;
}

export interface CreateSubjectBody {
  name: string;
  accentColor: string;
  plantType?: string;
}

export interface UpdateSubjectBody {
  name?: string;
  accentColor?: string;
}

export interface Plant {
  id: string;
  subjectId: string;
  subjectName: string;
  plantType: string;
  growthLevel: number;
  growthPoints: number;
  maxGrowthPoints: number;
  accentColor: string;
  createdAt: string;
}

export interface UpdatePlantBody {
  growthPoints?: number;
}

export type SessionSessionType =
  (typeof SessionSessionType)[keyof typeof SessionSessionType];

export const SessionSessionType = {
  routine: "routine",
  homework: "homework",
  deep_focus: "deep_focus",
} as const;

export type SessionState = (typeof SessionState)[keyof typeof SessionState];

export const SessionState = {
  initialized: "initialized",
  started: "started",
  warning: "warning",
  completed: "completed",
  aborted: "aborted",
} as const;

export interface Session {
  id: string;
  subjectId: string;
  subjectName: string;
  sessionType: SessionSessionType;
  state: SessionState;
  startTime: string;
  endTime?: string;
  durationMinutes: number;
  pointsEarned: number;
  createdAt: string;
}

export interface ActiveSessionResponse {
  session?: Session;
}

export type CreateSessionBodySessionType =
  (typeof CreateSessionBodySessionType)[keyof typeof CreateSessionBodySessionType];

export const CreateSessionBodySessionType = {
  routine: "routine",
  homework: "homework",
  deep_focus: "deep_focus",
} as const;

export interface CreateSessionBody {
  subjectId: string;
  sessionType: CreateSessionBodySessionType;
  durationMinutes: number;
}

export type UpdateSessionBodyState =
  (typeof UpdateSessionBodyState)[keyof typeof UpdateSessionBodyState];

export const UpdateSessionBodyState = {
  completed: "completed",
  aborted: "aborted",
} as const;

export interface UpdateSessionBody {
  state: UpdateSessionBodyState;
  endTime?: string;
}

export type SessionEventEventType =
  (typeof SessionEventEventType)[keyof typeof SessionEventEventType];

export const SessionEventEventType = {
  started: "started",
  warning: "warning",
  violation: "violation",
  completed: "completed",
  aborted: "aborted",
} as const;

export type SessionEventMetadata = { [key: string]: unknown };

export interface SessionEvent {
  id: string;
  sessionId: string;
  eventType: SessionEventEventType;
  eventTime: string;
  metadata?: SessionEventMetadata;
}

export type CreateSessionEventBodyEventType =
  (typeof CreateSessionEventBodyEventType)[keyof typeof CreateSessionEventBodyEventType];

export const CreateSessionEventBodyEventType = {
  started: "started",
  warning: "warning",
  violation: "violation",
  completed: "completed",
  aborted: "aborted",
} as const;

export type CreateSessionEventBodyMetadata = { [key: string]: unknown };

export interface CreateSessionEventBody {
  eventType: CreateSessionEventBodyEventType;
  metadata?: CreateSessionEventBodyMetadata;
}

export interface Wallet {
  balance: number;
  lastUpdated: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  reward: "reward",
  purchase: "purchase",
  bet: "bet",
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export interface Message {
  id: string;
  content: string;
  createdAt: string;
}

export interface CreateMessageBody {
  content: string;
  sessionId?: string;
}

export type InsightType = (typeof InsightType)[keyof typeof InsightType];

export const InsightType = {
  motivation: "motivation",
  warning: "warning",
  achievement: "achievement",
} as const;

export interface Insight {
  id: string;
  sessionId: string;
  type: InsightType;
  content: string;
  createdAt: string;
}

export interface UserStats {
  totalFocusMinutes: number;
  totalSessions: number;
  completedSessions: number;
  currentStreak: number;
  longestStreak: number;
  todayMinutes: number;
  weekMinutes: number;
}

export type GetSessionsParams = {
  limit?: number;
};