# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FocusOura API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: Get user subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a new subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubjectBody"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /insights/{sessionId}:
    get:
      operationId: getSessionInsights
      tags: [insights]
      summary: Get AI insights for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
    post:
      operationId: generateSessionInsights
      tags: [insights]
      summary: Generate AI insights for a completed session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Generated insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Subject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        accentColor:
          type: string
        plantId:
          type: string
        totalFocusMinutes:
          type: integer
        sessionCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - accentColor
        - totalFocusMinutes
        - sessionCount
        - createdAt
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        accentColor:
          type: string
        plantType:
          type: string
      required:
        - name
        - accentColor
    UpdateSubjectBody:
      type: object
      properties:
        name:
          type: string
        accentColor:
          type: string
...
    Insight:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        type:
          type: string
          enum: [motivation, warning, achievement]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - type
        - content
        - createdAt
...