# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusOura API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.string(),
  displayName: zod.string(),
  email: zod.string(),
  studyMode: zod.enum(["light", "night"]),
  notificationsEnabled: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Update current user
 */
export const UpdateMeBody = zod.object({
  displayName: zod.string().optional(),
  studyMode: zod.enum(["light", "night"]).optional(),
  notificationsEnabled: zod.boolean().optional(),
});

export const UpdateMeResponse = zod.object({
  id: zod.string(),
  displayName: zod.string(),
  email: zod.string(),
  studyMode: zod.enum(["light", "night"]),
  notificationsEnabled: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get user subjects
 */
export const GetSubjectsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  accentColor: zod.string(),
  plantId: zod.string().optional(),
  totalFocusMinutes: zod.number(),
  sessionCount: zod.number(),
  createdAt: zod.string(),
});
export const GetSubjectsResponse = zod.array(GetSubjectsResponseItem);

/**
 * @summary Create a new subject
 */
export const CreateSubjectBody = zod.object({
  name: zod.string(),
  accentColor: zod.string(),
  plantType: zod.string().optional(),
});

/**
 * @summary Update a subject
 */
export const UpdateSubjectParams = zod.object({
  subjectId: zod.coerce.string(),
});

export const UpdateSubjectBody = zod.object({
  name: zod.string().optional(),
  accentColor: zod.string().optional(),
});

export const UpdateSubjectResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  accentColor: zod.string(),
  plantId: zod.string().optional(),
  totalFocusMinutes: zod.number(),
  sessionCount: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a subject
 */
export const DeleteSubjectParams = zod.object({
  subjectId: zod.coerce.string(),
});

export const DeleteSubjectResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get user plants
 */
export const GetPlantsResponseItem = zod.object({
  id: zod.string(),
  subjectId: zod.string(),
  subjectName: zod.string(),
  plantType: zod.string(),
  growthLevel: zod.number(),
  growthPoints: zod.number(),
  maxGrowthPoints: zod.number(),
  accentColor: zod.string(),
  createdAt: zod.string(),
});
export const GetPlantsResponse = zod.array(GetPlantsResponseItem);

/**
 * @summary Update plant growth
 */
export const UpdatePlantParams = zod.object({
  plantId: zod.coerce.string(),
});

export const UpdatePlantBody = zod.object({
  growthPoints: zod.number().optional(),
});

export const UpdatePlantResponse = zod.object({
  id: zod.string(),
  subjectId: zod.string(),
  subjectName: zod.string(),
  plantType: zod.string(),
  growthLevel: zod.number(),
  growthPoints: zod.number(),
  maxGrowthPoints: zod.number(),
  accentColor: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get user sessions
 */
export const GetSessionsQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
});

export const GetSessionsResponseItem = zod.object({
  id: zod.string(),
  subjectId: zod.string(),
  subjectName: zod.string(),
  sessionType: zod.enum(["routine", "homework", "deep_focus"]),
  state: zod.enum([
    "initialized",
    "started",
    "warning",
    "completed",
    "aborted",
  ]),
  startTime: zod.string(),
  endTime: zod.string().optional(),
  durationMinutes: zod.number(),
  pointsEarned: zod.number(),
  createdAt: zod.string(),
});
export const GetSessionsResponse = zod.array(GetSessionsResponseItem);

/**
 * @summary Start a new focus session
 */
export const CreateSessionBody = zod.object({
  subjectId: zod.string(),
  sessionType: zod.enum(["routine", "homework", "deep_focus"]),
  durationMinutes: zod.number(),
});

/**
 * @summary Get currently active session
 */
export const GetActiveSessionResponse = zod.object({
  session: zod
    .object({
      id: zod.string(),
      subjectId: zod.string(),
      subjectName: zod.string(),
      sessionType: zod.enum(["routine", "homework", "deep_focus"]),
      state: zod.enum([
        "initialized",
        "started",
        "warning",
        "completed",
        "aborted",
      ]),
      startTime: zod.string(),
      endTime: zod.string().optional(),
      durationMinutes: zod.number(),
      pointsEarned: zod.number(),
      createdAt: zod.string(),
    })
    .optional(),
});

/**
 * @summary Update session state (complete or abort)
 */
export const UpdateSessionParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const UpdateSessionBody = zod.object({
  state: zod.enum(["completed", "aborted"]),
  endTime: zod.string().optional(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.string(),
  subjectId: zod.string(),
  subjectName: zod.string(),
  sessionType: zod.enum(["routine", "homework", "deep_focus"]),
  state: zod.enum([
    "initialized",
    "started",
    "warning",
    "completed",
    "aborted",
  ]),
  startTime: zod.string(),
  endTime: zod.string().optional(),
  durationMinutes: zod.number(),
  pointsEarned: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Record a session event
 */
export const CreateSessionEventParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const CreateSessionEventBody = zod.object({
  eventType: zod.enum([
    "started",
    "warning",
    "violation",
    "completed",
    "aborted",
  ]),
  metadata: zod.object({}).passthrough().optional(),
});

/**
 * @summary Get user wallet balance
 */
export const GetWalletResponse = zod.object({
  balance: zod.number(),
  lastUpdated: zod.string(),
});

/**
 * @summary Get user transactions
 */
export const GetTransactionsResponseItem = zod.object({
  id: zod.string(),
  type: zod.enum(["reward", "purchase", "bet"]),
  amount: zod.number(),
  description: zod.string(),
  createdAt: zod.string(),
});
export const GetTransactionsResponse = zod.array(GetTransactionsResponseItem);

/**
 * @summary Get a random anonymous motivational message
 */
export const GetRandomMessageResponse = zod.object({
  id: zod.string(),
  content: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Submit a motivational message after completing a session
 */
export const CreateMessageBody = zod.object({
  content: zod.string(),
  sessionId: zod.string().optional(),
});

/**
 * @summary Get AI insights for a session
 */
export const GetSessionInsightsParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetSessionInsightsResponseItem = zod.object({
  id: zod.string(),
  sessionId: zod.string(),
  type: zod.enum(["motivation", "warning", "achievement"]),
  content: zod.string(),
  createdAt: zod.string(),
});
export const GetSessionInsightsResponse = zod.array(
  GetSessionInsightsResponseItem,
);

/**
 * @summary Generate AI insights for a completed session
 */
export const GenerateSessionInsightsParams = zod.object({
  sessionId: zod.coerce.string(),
});

/**
 * @summary Get user focus statistics
 */
export const GetStatsResponse = zod.object({
  totalFocusMinutes: zod.number(),
  totalSessions: zod.number(),
  completedSessions: zod.number(),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  todayMinutes: zod.number(),
  weekMinutes: zod.number(),
});