# lib/api-zod/src/generated/types/updateSubjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FocusOura API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSubjectBody {
  name?: string;
  accentColor?: string;
}